# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KIU Admission & Career Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /career/paths:
    get:
      operationId: listCareerPaths
      tags: [career]
      summary: Get career paths for finalists based on their program
      parameters:
        - name: program
          in: query
          schema:
            type: string
        - name: faculty
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Career paths
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerPathList"
  /career/my-profile:
    get:
      operationId: getFinalistProfile
      tags: [career]
      summary: Get the current finalist's profile
      responses:
        "200":
          description: Finalist profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinalistProfile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertFinalistProfile
      tags: [career]
      summary: Create or update the current finalist's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinalistProfileRequest"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinalistProfile"
  /opportunities:
    get:
      operationId: listOpportunities
      tags: [opportunities]
      summary: List available job/internship opportunities
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [job, internship]
        - name: field
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of opportunities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunityList"
    post:
      operationId: createOpportunity
      tags: [opportunities]
      summary: Post a new opportunity (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpportunityRequest"
      responses:
        "201":
          description: Opportunity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
  /opportunities/{id}:
    get:
      operationId: getOpportunity
      tags: [opportunities]
      summary: Get an opportunity by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateOpportunity
      tags: [opportunities]
      summary: Update an opportunity (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpportunityRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
    delete:
      operationId: deleteOpportunity
      tags: [opportunities]
      summary: Delete an opportunity (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /opportunities/{id}/apply:
    post:
      operationId: applyForOpportunity
      tags: [opportunities]
      summary: Apply for a job/internship opportunity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpportunityApplicationRequest"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunityApplication"
        "409":
          description: Already applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /opportunities/applications/{id}:
    patch:
      operationId: updateOpportunityApplicationStatus
      tags: [opportunities]
      summary: Update opportunity application status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOpportunityApplicationStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunityApplication"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Program:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        faculty:
          type: string
        department:
          type: string
        level:
          type: string
          enum: [diploma, degree]
        duration:
          type: string
        description:
          type: string
        entryRequirements:
          type: string
        minOlevelPoints:
          type: integer
        minAlevelPoints:
          type: integer
        availableSlots:
          type: integer
      required:
        - id
        - name
        - code
        - faculty
        - level
...
    CareerPath:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        relatedPrograms:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        potentialRoles:
          type: array
          items:
            type: string
        averageSalaryRange:
          type: string
        growthOutlook:
          type: string
        industryField:
          type: string
      required:
        - id
        - title
        - description
        - relatedPrograms
        - skills
        - potentialRoles
        - industryField
    CareerPathList:
      type: object
      properties:
        careerPaths:
          type: array
          items:
            $ref: "#/components/schemas/CareerPath"
        total:
          type: integer
      required:
        - careerPaths
        - total
    FinalistProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        programId:
          type: integer
        program:
          $ref: "#/components/schemas/Program"
        studentNumber:
          type: string
        yearOfStudy:
          type: integer
        graduationYear:
          type: integer
        gpa:
          type: number
        skills:
          type: array
          items:
            type: string
        bio:
          type: string
        linkedinUrl:
          type: string
        cvUrl:
          type: string
        isFinalist:
          type: boolean
      required:
        - id
        - userId
        - programId
        - studentNumber
        - yearOfStudy
        - isFinalist
    FinalistProfileRequest:
      type: object
      properties:
        programId:
          type: integer
        studentNumber:
          type: string
        yearOfStudy:
          type: integer
        graduationYear:
          type: integer
        gpa:
          type: number
        skills:
          type: array
          items:
            type: string
        bio:
          type: string
        linkedinUrl:
          type: string
        cvUrl:
          type: string
      required:
        - programId
        - studentNumber
        - yearOfStudy
    Opportunity:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        organization:
          type: string
        type:
          type: string
          enum: [job, internship]
        description:
          type: string
        requirements:
          type: string
        requiredPrograms:
          type: array
          items:
            type: string
        requiredSkills:
          type: array
          items:
            type: string
        location:
          type: string
        salaryRange:
          type: string
        applicationDeadline:
          type: string
          format: date
        contactEmail:
          type: string
        isActive:
          type: boolean
        postedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        applicantCount:
          type: integer
      required:
        - id
        - title
        - organization
        - type
        - description
        - requirements
        - applicationDeadline
        - isActive
    OpportunityList:
      type: object
      properties:
        opportunities:
          type: array
          items:
            $ref: "#/components/schemas/Opportunity"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - opportunities
        - total
        - page
        - limit
    OpportunityRequest:
      type: object
      properties:
        title:
          type: string
        organization:
          type: string
        type:
          type: string
          enum: [job, internship]
        description:
          type: string
        requirements:
          type: string
        requiredPrograms:
          type: array
          items:
            type: string
        requiredSkills:
          type: array
          items:
            type: string
        location:
          type: string
        salaryRange:
          type: string
        applicationDeadline:
          type: string
          format: date
        contactEmail:
          type: string
        isActive:
          type: boolean
      required:
        - title
        - organization
        - type
        - description
        - requirements
        - applicationDeadline
    OpportunityApplicationRequest:
      type: object
      properties:
        coverLetter:
          type: string
        cvUrl:
          type: string
        additionalInfo:
          type: string
      required:
        - coverLetter
    OpportunityApplication:
      type: object
      properties:
        id:
          type: integer
        opportunityId:
          type: integer
        opportunity:
          $ref: "#/components/schemas/Opportunity"
        userId:
          type: integer
        status:
          type: string
          enum: [applied, shortlisted, interview_scheduled, accepted, rejected]
        coverLetter:
          type: string
        cvUrl:
          type: string
        additionalInfo:
          type: string
        adminNotes:
          type: string
        appliedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        applicantName:
          type: string
        applicantEmail:
          type: string
      required:
        - id
        - opportunityId
        - userId
        - status
        - coverLetter
...
    UpdateOpportunityApplicationStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [applied, shortlisted, interview_scheduled, accepted, rejected]
        adminNotes:
          type: string
      required:
        - status