# lib/api-zod/src/generated/types/opportunityApplication.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KIU Admission & Career Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Opportunity } from "./opportunity";
import type { OpportunityApplicationStatus } from "./opportunityApplicationStatus";

export interface OpportunityApplication {
  id: number;
  opportunityId: number;
  opportunity?: Opportunity;
  userId: number;
  status: OpportunityApplicationStatus;
  coverLetter: string;
  cvUrl?: string;
  additionalInfo?: string;
  adminNotes?: string;
  appliedAt?: Date;
  updatedAt?: Date;
  applicantName?: string;
  applicantEmail?: string;
}